fontrenderer: remove '\n' handling
'\n' in a text was not working correctly - it did not reset the
X position correctly and it did jump out of the assigned box.
Visible errors were e.g. on epgplus, 13.0E SF1, where the short
description often contains '\n' and thus garbage was painted
below the epgplus window (and did not get cleared afterwards).
For now we replace '\n' with '¶' (0x00b6). Later, if no offenders
are found, it should simply be removed.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@872 e54a6e83-5905-42d5-8d5c-058d10e6a962