abstime.c: add 7 day offset to time_monotonic() functions
The start point of the monotonic clock is undefined, but in practice
it is often zero. There is lots of code that does not assume that a
timestamp can be that low and then fails in subtle ways.
Add a 7 day offset to the clock value to work around that.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1271 e54a6e83-5905-42d5-8d5c-058d10e6a962