neutrino: fix recording from standby
After the parental PIN fixes, zapit did no longer zap on leaving standby
in order to not bypass the parental PIN code. Unfortunately this broke
recording from soft standby if the channel to be recorded was the same
as the live channel before standby, since on wakeup it was not tuned.
Fix this by tuning the channel on zapit:leaveStandby(), but without
starting playback. Playback start is triggered by neutrino including the
PIN check.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1314 e54a6e83-5905-42d5-8d5c-058d10e6a962