dboxinfo: simplify code for FS/memory size display
Before, the units to display for filesystems / memory was based solely
on the filesystem type. This would lead to e.g. a 16MB FAT partition
being displayed as "0.01G" which is not too useful. Additionally, the
code made the distinction between RAM and FS in many places. Simplify
the code by determining the total/free/used in bytes, then pass that
number to a conversion function which returns a string with suffix.
Additional benefit: avoids use of float, code size significantly
smaller.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-experimental@1375 e54a6e83-5905-42d5-8d5c-058d10e6a962