*neutrino record: reduce multiple code for record calls
Some lines with recording-actions calls in record.cpp and neutrino.cpp were redundant.
Now we use only the exec() call with action key "Record" from record.cpp.
So we use the same functions with record button and with the recording menu item
in  user menu. This also causes, that we can see the infobar as response for user
on used menu item in user menu.

git-svn-id: file:///home/bas/coolstream_public_svn/THIRDPARTY/applications/neutrino-beta@1803 e54a6e83-5905-42d5-8d5c-058d10e6a962