spark: blit framebuffer once per second
Just for the case that some code has not yet added the correct
fb->blit() calls, call it once per second to make sure that
stuff appears on the display.
If you find menus or stuff which is lagging heavily, they are
probably missing those blit() calls.
Of course this has nothing to do with the LED display code,
but putting it there saves us from using yet another thread.