CComponentsItem: add members getItemType(), cc_item_type
Small helper for detection of item types.
This could be useful, if it is necessary, to track these at runtime or
debugging, because it's not really detectable, which items just be used.
e.g forms could contain very much items, but many of these items
are inherited from CComponentsItem and an assignment to a certain subclass.