CTimeOSD: use class konform member names
CTimeOSD is inherited from CComponentsForm/Item.
Therefore, it is usual to keep the functional layout.
Overwritable virtual members should be used with same
functionality. kill() is similar with paintBackground() known from
CFrameBuffer, hide(bool) do restore saved background and expects a
parameter.

TODO: It works, but class layout of CTimeOSD is currently not conform
with Components, because the functionalities are partially undermined
and must be reworked.
For Example: timescale and clock object not used as sub objects
in a parent form object.