locales: rename plugin entries for Lua-Plugins
Maybe it's sensibly to use no explicit description for Lua-plugins.
The user could unnecessarily ask "What is this?".
"Lua" is more a subject-specific description or a technical term.
Therefore for users it should be reasonable to use "Plugins" as general
description.

TODO: use of these terms in the user menus, in particular "plugin types"