Main Menu: try to add dynamic possibility to disable items on runtime
The issue is that it's not really possible to disable/enable menu items on
runtime with an existant menu widget instance eg with personalized menu items.

Here we need a dynamic solution to disable items depends on stb-mode (mode_ts, mode_tv etc)
This should be solved here with an additional parameter for personalized items.

New paramter is named: disable_condition
Possible alvailable values at the moment are:

  DCOND_MODE_NONE as default

  DCOND_MODE_TV
  DCOND_MODE_RADIO
  DCOND_MODE_TS

includes some improvements by Sven

- menue: remove old_iconName handling
  ... icons should be painted on deactivated items too

- menue: try to fix position <-> selection missmatch