CInfoViewer: ensure removing of timescale bar after killed infobar
In some constellations it was possible that timescale  was not removed
from screen after closing of infoviewer.
e.g. user switched from a
channel with epg data to a channel without epg data and infobar was closed
before epg data was complete arrived.

btw: it seems INFOBAR_PROGRESSBAR_ARRANGEMENT_DEFAULT
is not not really "default"