infoviewer: change show_Data() semantics slightly
the "calledFromEvent" parameter was only used to force display of "no
EPG available". It was set to true when calling from an "zap complete"
or "epg event received" event, which is good, but also when calling from
the display update timer, which is "strange".
Now setting this parameter to false forces the update of the current
event (if available) which is done from display update timer.
This fixes the update of the remaining time display with constantly
displayed infobar (radio mode).