CMsgBox: ensure repaint of background
This fixes "transparent holes", because textbox object is responsible
for text.

The problem here is that the message box has a fixed width and text box
width is not really considered. This should be fixed, but CMsgBox class
is used only by CMovieinfo class and this should be enough for the moment.