CThemes: try to fix possible artefacts after changed theme
This ensures a clean screen before repaint after changed theme.

In some constellations (eg. with color gradient) it was possible to
see some old artefacts inside widget items, eg. in header.
This should avoid this.
Small disadvantage: screen needs complete repaint after changed theme, but
this should be not a problem in this case.