opkg_manager: use system-update script if available
opkg has several problems with complex system updates, especially when
having to update itself and with the package ordering during update
To avoid hard coding the workarounds into neutrino, just call a script
named "system-update" (if available) instead of "opkg upgrade".
This script can be tuned to the specific target and its problems.