start modularizing framebuffer acceleration architecture
This is just copied framebuffer.{h,cpp} => fb_generic.{h,cpp}
The idea is to implement CFrameBuffer as generic framebuffer class that
does work on almost every hardware without acceleration and / or
hardware dependencies.
CFbAccel will be a derived class (and different on many architectures)
which implements the accelerated functions.
This does not yet compile, verbatim copies to track development history.