tuxtxt: convert to int-pointer arithmetic, work with all framebuffers
This converts the drawing code from byte-pointer to fb_pixel_t-pointer
arithmetic, making the calculations more obvious and probably more
efficient.
Additionally, the color tables are now generated in a way that it should
work regardless of the frame buffer color format. Allows to remove a few
hardware specific #ifdefs.