CComponentsItem: fix possible artefacts if fb box is smaller the radius
If box height/2 is smaller than corner radius it is possible
we have artefacts after killed or hided fb box. This effect only I have seen
with full hd resolution, in hd ready mode it was ok.
Sideeffects with this changes I have not seen, but I can't exclude. Must be watched.