CAudioPlayerGui: add cc scquare objekt as title box, fix caption bg colors
Use of cc square object saves unnecessary paintBoxRel() calls.
BgColors of metatdata and time display were different to
titlebox body color, but was not to see with all themes.

btw: time box was too much on the right side. Frame of titelbox
was overpainted but was not very noticeable if frame width < 2.