CComponentsHeader: rework and simplified logo handling
This should fix some issues with possible overlapps
of logos with title. With some Conditions with text
alignment on the right side needs a fix.
There is an issue with size and position of text object, but is stll not
used at the moment. For all other conditions of logo and text alignments now
it should be not a problem (I hope) and have more code clarity.