Colorsettings: add possibility to change shadow color
Shadow color was bound to infobar color and it's not really
understandable why, because with some colors eg. black, the shadow is
grey or with other colors it is exaggerated colored and seems to be
not realistically and looks unusual if shadow is brighter or colorful
than foreground. Some themes will thank it.

TODO: update theme files with matching values if required