- infoviewer: rework tuner icons
 Conflicts:
	data/icons/status/channel/tuner_1.png
	data/icons/status/channel/tuner_2.png
	data/icons/status/channel/tuner_3.png
	data/icons/status/channel/tuner_4.png

Signed-off-by: Thilo Graf <dbt@novatux.de>

I think it's not an excellent idea to have many icons for all possible
tuner numbers. We should use one and only icon without any number
and paint a text over the icon. This should be possible with a simple
cc text object that contains a defined png-file file as background
and a digit as text.