partial revert2: "- remove --enable-testing switch; testmenu is intergrated into usermenu"
The entry for the test menu at the end of the list inside the main menu
was implemented not without reason and not for fun.
Moving the test menu to the user menus (though it is not intended
for users!), complicates the access to the test menu without prior
configuration with nice shortcuts.
Searching for a colored key or any other key (no matter it
is configured or not)  is always more awkward than navigation with
the control cross keys, which can be used blind too.
If you need it often, that's just annoying.