Revert "- sectionsd: replace hardcoded timestamp with BUILT_TIMESTAMP"
This reverts commit 521239bb2b560a69f328eab4f1708a583ab83d37.

Depending on the build system and preferences of build system maintainers
the file "version.h" is not absolutely necessary or could be empty and
if this file is not exists, build will break.
This needs another solution.
Why not solve it at runtime? That would at least be more sustainable,
because long-term the timestamp would not age and recompile could be avoided.