- neutrino.cpp: stop lcd4l thread for flashing and remove dead code
In src/system/flashtool.cpp there was still old, unused code,
which once stopped LCD4L during flash. However, it has not been
jumped and executed for a long time. I have removed it now and
introduced the corresponding code section in neutrino.cpp in
stop_deamons() and reactivated so the stopping of the daemon.
The sighandler() function calls up also stop_daemons(),
so I removed the code there