Remove dead code from CRCInput::getMsg_us()
The two events in question are posted in CInfoViewer::getEPG()
and they're handled in CRemoteControl::handleMsg().

They are written to the high/low priority pipes only, so there should
be no way that they end up in the event socket, which is what the code
path in question here deals with. Apart from that, the type the buffer
was casted to was completely wrong, so if this code was ever executed,
it would have faulted.