widget/menu: add member 'type_name' to CMenuItem class
This makes it easier to distinguish the object type. Normally,
type queries would be done using typeid-operation or dynamic_cast,
but we usually have RTTI disabled in the Makefiles. Presumably,
this is to avoid unnecessarily increasing the size of the binaries.
The type information defined by this member is currently only defined
in the CMenuItem class, so all subclasses are also equipped with it.
An exception is the CMenuSeparator class. If further definitions are
necessary in the subclasses, they can be retrofitted as needed.