system/helpers: Add new Parameter 'start_pos' to str_replace()
This allows flexibility to specify a starting position from where to search
for the first occurrence of the search string. By default,
the position is set to 0 and the function will behave as before.
However, when a position is provided,
the function will replace only the first occurrence of
the search string after that position.